Fixed some of the tests to allow it to identify which two files were causing a merge conflict.
